Testing POST mapping

+++ src/main/java/com/example/bugyourspot/reservation/Reservation.java
@@ -19,8 +19,8 @@ public class Reservation {
     )
 
     private Long reservationId;
-    private Integer clientId;
-    private Integer customerId;
+    private Long clientId;
+    private Long customerId;
     private LocalDateTime startTime;
     private Integer numSlots;
 
@@ -29,7 +29,7 @@ public class Reservation {
     public Reservation(){
     }
 
-    public Reservation(Long reservationId, Integer clientId, Integer customerId, LocalDateTime startTime, Integer numSlots) {
+    public Reservation(Long reservationId, Long clientId, Long customerId, LocalDateTime startTime, Integer numSlots) {
         this.reservationId = reservationId;
         this.customerId = customerId;
         this.clientId = clientId;
@@ -37,18 +37,18 @@ public class Reservation {
         this.numSlots = numSlots;
     }
 
-    public Reservation(Integer clientId, Integer customerId, LocalDateTime startTime, Integer numSlots) {
+    public Reservation(Long clientId, Long customerId, LocalDateTime startTime, Integer numSlots) {
         this.clientId = clientId;
         this.customerId = customerId;
         this.startTime = startTime;
         this.numSlots = numSlots;
     }
 
-    public Integer getClientId() {
+    public Long getClientId() {
         return clientId;
     }
 
-    public Integer getCustomerId() { return customerId; }
+    public Long getCustomerId() { return customerId; }
 
     public LocalDateTime getStartTime() {
         return startTime;
@@ -60,11 +60,11 @@ public class Reservation {
 
     public Long getReservationId() { return reservationId; }
 
-    public void setClientId(Integer clientId) {
+    public void setClientId(Long clientId) {
         this.clientId = clientId;
     }
 
-    public void setCustomerId(Integer customerId) { this.customerId = customerId; }
+    public void setCustomerId(Long customerId) { this.customerId = customerId; }
 
     public void setStartTime(LocalDateTime startTime) {
         this.startTime = startTime;
